[ui]#152 수정 뷰 라이팅 워싱 완료 및 refactor
