+++ app/src/main/java/kr/co/nottodo/presentation/modification/viewmodel/ModificationViewModel.kt
@@ -7,7 +7,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import kr.co.nottodo.R
 import kr.co.nottodo.data.remote.api.ServicePool
 import kr.co.nottodo.data.remote.api.ServicePool.modificationService
 import kr.co.nottodo.data.remote.model.ResponseRecentMissionListDto
@@ -15,11 +14,9 @@ import kr.co.nottodo.data.remote.model.ResponseRecommendSituationListDto
 import kr.co.nottodo.data.remote.model.modification.RequestModificationDto
 import kr.co.nottodo.data.remote.model.modification.ResponseGetMissionDates
 import kr.co.nottodo.data.remote.model.modification.ResponseModificationDto
-import kr.co.nottodo.presentation.modification.view.ModificationActivity.Companion.NotTodoData
-import kr.co.nottodo.util.NotTodoAmplitude.trackEventWithProperty
+import kr.co.nottodo.presentation.modification.model.NotTodoData
 import kr.co.nottodo.util.getErrorMessage
 import kr.co.nottodo.view.calendar.monthly.util.achievementConvertStringToDate
-import kr.co.nottodo.view.calendar.monthly.util.convertDateStringToInt
 import kr.co.nottodo.view.calendar.monthly.util.isToday
 import kr.co.nottodo.view.calendar.monthly.util.isTomorrow
