+++ app/src/main/res/layout/activity_addition.xml
@@ -769,7 +769,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="8dp"
-                            android:text="낫투두 - 유튜브 보지않기"
+                            android:text="@string/ex_mission_youtube"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/bg_f2f2f7"
                             app:layout_constraintStart_toEndOf="@id/tv_addition_action_example"
@@ -782,7 +782,7 @@
                             android:layout_marginTop="8dp"
                             android:layout_marginBottom="37dp"
                             android:lineSpacingExtra="5dp"
-                            android:text="실천 방법 - 유튜브 프리미엄 구독 취소\n실천 방법 - 공부 중에 가방 안에 핸드폰 넣기"
+                            android:text="@string/ex_action_youtube"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintBottom_toBottomOf="parent"
@@ -952,7 +952,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="5dp"
-                            android:text="운동 중 20분 이상 쉬지 않기"
+                            android:text="@string/ex_mission_work_out"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintStart_toEndOf="@id/tv_addition_goal_example_mission_title"
@@ -967,7 +967,7 @@
                             android:background="@drawable/rectangle_solid_gray_2_radius_50"
                             android:paddingHorizontal="13dp"
                             android:paddingVertical="3dp"
-                            android:text="목표"
+                            android:text="@string/goal"
                             android:textAppearance="@style/R12"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintBottom_toBottomOf="parent"
@@ -979,7 +979,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="5dp"
-                            android:text="근손실 막고 근육량 늘리기"
+                            android:text="@string/ex_goal_work_out"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintStart_toEndOf="@id/tv_addition_goal_example_goal_title"
