+++ app/src/main/res/layout/activity_modification.xml
@@ -294,12 +294,14 @@
                             android:layout_height="wrap_content"
                             android:layout_marginTop="25dp"
                             android:backgroundTint="@color/gray_3_5d5d6b"
+                            android:hint="낫투두는 '~않기'로 작성하면 더 쉬워요."
                             android:imeOptions="actionDone"
                             android:inputType="text"
                             android:maxLength="20"
                             android:text="@={vm.mission}"
                             android:textAppearance="@style/M14"
                             android:textColor="@color/white"
+                            android:textColorHint="@color/gray_4_9398aa"
                             app:layout_constraintEnd_toEndOf="parent"
                             app:layout_constraintStart_toStartOf="parent"
                             app:layout_constraintTop_toBottomOf="@id/tv_modification_mission_opened_desc" />
@@ -465,7 +467,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginTop="28dp"
-                            android:text="추천 키워드"
+                            android:text="추천 상황"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/white"
                             app:layout_constraintStart_toStartOf="parent"
@@ -522,6 +524,7 @@
                             android:layout_height="wrap_content"
                             android:layout_marginVertical="18dp"
                             android:layout_marginStart="82dp"
+                            android:lineSpacingExtra="9dp"
                             android:text="@string/addition_input"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_3_5d5d6b"
@@ -544,11 +547,12 @@
                             android:id="@+id/iv_modification_action_closed_check"
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
+                            android:layout_marginTop="16dp"
                             android:layout_marginEnd="16dp"
                             android:src="@drawable/ic_check"
                             android:visibility="gone"
                             app:layout_constraintEnd_toEndOf="parent"
-                            app:layout_constraintTop_toTopOf="@id/tv_modification_action_closed_input" />
+                            app:layout_constraintTop_toTopOf="parent" />
 
                     </androidx.constraintlayout.widget.ConstraintLayout>
 
@@ -717,7 +721,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="8dp"
-                            android:text="유튜브 보지않기"
+                            android:text="@string/ex_mission_youtube"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/bg_f2f2f7"
                             app:layout_constraintStart_toEndOf="@id/tv_modification_action_example"
@@ -729,7 +733,8 @@
                             android:layout_height="wrap_content"
                             android:layout_marginTop="8dp"
                             android:layout_marginBottom="37dp"
-                            android:text="유튜브 프리미엄 구독 취소\n공부 중에 가방 안에 핸드폰 넣기"
+                            android:lineSpacingExtra="5dp"
+                            android:text="@string/ex_action_youtube"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintBottom_toBottomOf="parent"
@@ -899,7 +904,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="5dp"
-                            android:text="유튜브 보지 않기"
+                            android:text="@string/ex_mission_work_out"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintStart_toEndOf="@id/tv_modification_goal_example_mission_title"
@@ -926,7 +931,7 @@
                             android:layout_width="wrap_content"
                             android:layout_height="wrap_content"
                             android:layout_marginStart="5dp"
-                            android:text="유튜브 프리미엄 구독 취소"
+                            android:text="@string/ex_goal_work_out"
                             android:textAppearance="@style/R14"
                             android:textColor="@color/gray_4_9398aa"
                             app:layout_constraintStart_toEndOf="@id/tv_modification_goal_example_goal_title"
